# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AiProviderStatus {
  gemini: boolean;
  openrouter: boolean;
}

export type AiSuggestionInputProvider = typeof AiSuggestionInputProvider[keyof typeof AiSuggestionInputProvider];


export const AiSuggestionInputProvider = {
  gemini: 'gemini',
  openrouter: 'openrouter',
} as const;

export interface AiSuggestionInput {
  provider: AiSuggestionInputProvider;
  /**
     * @minLength 1
     * @maxLength 12000
     */
  script: string;
  /** @minLength 1 */
  consentToken: string;
}

export type AiConsentInputProvider = typeof AiConsentInputProvider[keyof typeof AiConsentInputProvider];


export const AiConsentInputProvider = {
  gemini: 'gemini',
  openrouter: 'openrouter',
} as const;

export interface AiConsentInput {
  provider: AiConsentInputProvider;
  /**
     * @minLength 1
     * @maxLength 12000
     */
  script: string;
  reviewed: true;
}

export interface AiConsentResult {
  consentToken: string;
  expiresInSeconds: number;
}

export type AiSuggestionResultProvider = typeof AiSuggestionResultProvider[keyof typeof AiSuggestionResultProvider];


export const AiSuggestionResultProvider = {
  gemini: 'gemini',
  openrouter: 'openrouter',
} as const;

export type AiSuggestionResultUsage = {
  used: number;
  limit: number;
  remaining: number;
};

export interface AiSuggestionResult {
  provider: AiSuggestionResultProvider;
  /** @maxItems 5 */
  hooks: string[];
  /** @maxItems 8 */
  broll: string[];
  fallback?: boolean;
  message?: string;
  usage: AiSuggestionResultUsage;
}