# lib/api-zod/src/generated/types/aiConsentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiConsentInputProvider } from './aiConsentInputProvider';

export interface AiConsentInput {
  provider: AiConsentInputProvider;
  /**
     * @minLength 1
     * @maxLength 12000
     */
  script: string;
  reviewed: true;
}