# lib/api-zod/src/generated/types/aiSuggestionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiSuggestionInputProvider } from './aiSuggestionInputProvider';

export interface AiSuggestionInput {
  provider: AiSuggestionInputProvider;
  /**
     * @minLength 1
     * @maxLength 12000
     */
  script: string;
  /** @minLength 1 */
  consentToken: string;
}