# lib/api-zod/src/generated/types/aiSuggestionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiSuggestionResultProvider } from './aiSuggestionResultProvider';
import type { AiSuggestionResultUsage } from './aiSuggestionResultUsage';

export interface AiSuggestionResult {
  provider: AiSuggestionResultProvider;
  /** @maxItems 5 */
  hooks: string[];
  /** @maxItems 8 */
  broll: string[];
  fallback?: boolean;
  message?: string;
  usage: AiSuggestionResultUsage;
}